Add files via upload
Hey guys, 
In this jupyter notebook i have successfully automated a function which helps in plotting and saving graphs of all the columns in the data set. During EDA it's essential to look at graphs of all columns and doing it manually is tiresome. With this function you can plot and save all graphs in your chosen directory.